Models now provide getContentMenuActions
