+++ BabelWiresQtUi/ComplexValueEditors/MapEditor/mapEditor.cpp
@@ -10,23 +10,24 @@
 #include <BabelWiresQtUi/ComplexValueEditors/MapEditor/mapModel.hpp>
 #include <BabelWiresQtUi/ComplexValueEditors/MapEditor/mapModelDelegate.hpp>
 #include <BabelWiresQtUi/ComplexValueEditors/MapEditor/typeWidget.hpp>
+#include <BabelWiresQtUi/ContextMenu/contextMenu.hpp>
 #include <BabelWiresQtUi/ModelBridge/accessModelScope.hpp>
 #include <BabelWiresQtUi/ModelBridge/projectBridge.hpp>
 #include <BabelWiresQtUi/uiProjectContext.hpp>
 
-#include <BabelWiresLib/ValueTree/modelExceptions.hpp>
-#include <BabelWiresLib/Types/Map/mapType.hpp>
+#include <BabelWiresLib/Project/Commands/addModifierCommand.hpp>
+#include <BabelWiresLib/Project/Modifiers/modifier.hpp>
+#include <BabelWiresLib/Project/Modifiers/valueAssignmentData.hpp>
+#include <BabelWiresLib/Project/Nodes/node.hpp>
+#include <BabelWiresLib/ProjectExtra/projectDataLocation.hpp>
 #include <BabelWiresLib/Types/Map/Commands/setMapCommand.hpp>
 #include <BabelWiresLib/Types/Map/Commands/setMapSourceTypeCommand.hpp>
 #include <BabelWiresLib/Types/Map/Commands/setMapTargetTypeCommand.hpp>
 #include <BabelWiresLib/Types/Map/MapProject/mapSerialization.hpp>
-#include <BabelWiresLib/Project/Commands/addModifierCommand.hpp>
-#include <BabelWiresLib/Project/Nodes/node.hpp>
-#include <BabelWiresLib/Project/Modifiers/valueAssignmentData.hpp>
-#include <BabelWiresLib/Project/Modifiers/modifier.hpp>
 #include <BabelWiresLib/Types/Map/SumOfMaps/sumOfMapsType.hpp>
+#include <BabelWiresLib/Types/Map/mapType.hpp>
+#include <BabelWiresLib/ValueTree/modelExceptions.hpp>
 #include <BabelWiresLib/ValueTree/valueTreeNode.hpp>
-#include <BabelWiresLib/ProjectExtra/projectDataLocation.hpp>
 
 #include <QDialogButtonBox>
 #include <QFileDialog>
@@ -91,13 +92,17 @@ babelwires::MapEditor::MapEditor(QWidget* parent, ProjectBridge& projectBridge, 
             m_typeRef = mapTreeNode.getTypeRef();
             const MapValue& mapValue = getMapValueFromProject(scope);
             if (mapTreeNode.getType().as<MapType>()) {
-                m_map.setAllowedSourceTypeRefs(MapProject::AllowedTypes{{mapTreeNode.getType().is<MapType>().getSourceTypeRef()}});
-                m_map.setAllowedTargetTypeRefs(MapProject::AllowedTypes{{mapTreeNode.getType().is<MapType>().getTargetTypeRef()}});
+                m_map.setAllowedSourceTypeRefs(
+                    MapProject::AllowedTypes{{mapTreeNode.getType().is<MapType>().getSourceTypeRef()}});
+                m_map.setAllowedTargetTypeRefs(
+                    MapProject::AllowedTypes{{mapTreeNode.getType().is<MapType>().getTargetTypeRef()}});
             } else {
-                const SumOfMapsType *const sumOfMaps = mapTreeNode.getType().as<SumOfMapsType>();
+                const SumOfMapsType* const sumOfMaps = mapTreeNode.getType().as<SumOfMapsType>();
                 assert(sumOfMaps && "MapEditor expecting a MapType of SumOfMapsType");
-                m_map.setAllowedSourceTypeRefs(MapProject::AllowedTypes{sumOfMaps->getSourceTypes(), sumOfMaps->getIndexOfDefaultSourceType()});
-                m_map.setAllowedTargetTypeRefs(MapProject::AllowedTypes{sumOfMaps->getTargetTypes(), sumOfMaps->getIndexOfDefaultTargetType()});
+                m_map.setAllowedSourceTypeRefs(
+                    MapProject::AllowedTypes{sumOfMaps->getSourceTypes(), sumOfMaps->getIndexOfDefaultSourceType()});
+                m_map.setAllowedTargetTypeRefs(
+                    MapProject::AllowedTypes{sumOfMaps->getTargetTypes(), sumOfMaps->getIndexOfDefaultTargetType()});
             }
             m_map.setMapValue(mapValue);
             {
@@ -227,7 +232,8 @@ babelwires::MapEditor::tryGetMapValueAssignmentData(AccessModelScope& scope) con
     return modifier->getModifierData().as<ValueAssignmentData>();
 }
 
-babelwires::ValueHolderTemplate<babelwires::MapValue> babelwires::MapEditor::tryGetMapValueFromProject(AccessModelScope& scope) const {
+babelwires::ValueHolderTemplate<babelwires::MapValue>
+babelwires::MapEditor::tryGetMapValueFromProject(AccessModelScope& scope) const {
     if (const ValueAssignmentData* const modifier = tryGetMapValueAssignmentData(scope)) {
         if (ValueHolderTemplate<MapValue> mapValue = modifier->getValue().asValueHolder<MapValue>()) {
             return mapValue;
@@ -247,8 +253,7 @@ void babelwires::MapEditor::updateMapFromProject() {
     ValueHolderTemplate<MapValue> mapValueFromProject = tryGetMapValueFromProject(scope);
     if (mapValueFromProject) {
         getUserLogger().logInfo() << "Refreshing the map from the project";
-        executeCommand(
-            std::make_unique<SetMapCommand>("Refresh the map from the project", mapValueFromProject));
+        executeCommand(std::make_unique<SetMapCommand>("Refresh the map from the project", mapValueFromProject));
     } else {
         warnThatMapNoLongerInProject("Cannot refresh the map.");
     }
@@ -349,8 +354,13 @@ void babelwires::MapEditor::warnThatMapNoLongerInProject(const std::string& oper
 
 void babelwires::MapEditor::onCustomContextMenuRequested(const QPoint& pos) {
     QModelIndex index = m_mapView->indexAt(pos);
-    QMenu* const menu = m_mapModel->getContextMenu(index);
-    if (menu) {
+    std::vector<ContextMenuEntry> actions;
+    m_mapModel->getContextMenuActions(actions, index);
+    if (!actions.empty()) {
+        ContextMenu* menu = new ContextMenu(*m_mapModel, index);
+        for (auto&& action : actions) {
+            menu->addContextMenuEntry(std::move(action));
+        }
         menu->popup(m_mapView->mapToGlobal(pos));
     }
 }
