+++ BabelWiresQtUi/ContextMenu/contextMenu.hpp
@@ -15,25 +15,6 @@
 #include <QAbstractItemModel>
 
 namespace babelwires {
-    /// Holds a group of actions, which will be added to the menu in a QActionGroup.
-    /// By default, the actions will be exclusive.
-    class ContextMenuGroup {
-      public:
-        ContextMenuGroup(QString name,
-                                QActionGroup::ExclusionPolicy policy = QActionGroup::ExclusionPolicy::Exclusive)
-            : m_groupName(name)
-            , m_exclusionPolicy(policy) {}
-
-        void addContextMenuAction(std::unique_ptr<ContextMenuAction> action);
-
-        QActionGroup::ExclusionPolicy m_exclusionPolicy;
-        QString m_groupName;
-        std::vector<std::unique_ptr<ContextMenuAction>> m_actions;
-    };
-
-    using ContextMenuEntry =
-        std::variant<std::unique_ptr<ContextMenuAction>, std::unique_ptr<ContextMenuGroup>>;
-
     /// The pop-up context menu used for the elements of an QAbstractItemModel.
     class ContextMenu : public QMenu {
         Q_OBJECT
