+++ BabelWiresQtUi/ContextMenu/contextMenuAction.hpp
@@ -26,4 +26,24 @@ namespace babelwires {
         /// Dispatches to actionTriggered, obtaining the arguments from the parent menu.
         void onTriggeredFired();
     };
+
+        /// Holds a group of actions, which will be added to the menu in a QActionGroup.
+    /// By default, the actions will be exclusive.
+    class ContextMenuGroup {
+      public:
+        ContextMenuGroup(QString name,
+                                QActionGroup::ExclusionPolicy policy = QActionGroup::ExclusionPolicy::Exclusive)
+            : m_groupName(name)
+            , m_exclusionPolicy(policy) {}
+
+        void addContextMenuAction(std::unique_ptr<ContextMenuAction> action);
+
+        QActionGroup::ExclusionPolicy m_exclusionPolicy;
+        QString m_groupName;
+        std::vector<std::unique_ptr<ContextMenuAction>> m_actions;
+    };
+
+    using ContextMenuEntry =
+        std::variant<std::unique_ptr<ContextMenuAction>, std::unique_ptr<ContextMenuGroup>>;
+
 } // namespace babelwires
